Added weather to header
