+++ src/app/dashboard/dashboard.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { DashboardComponent } from './dashboard.component';
-import { WeatherDisplayComponent } from './weather-display/weather-display.component';
+import { WeatherDisplayComponent } from '../generated-list-wrapper/weather-display/weather-display.component';
 import { FilterTypeComponent } from './filters/filter-type/filter-type.component';
 import { FilterGenderComponent } from './filters/filter-gender/filter-gender.component';
 import { FilterDurationComponent } from './filters/filter-duration/filter-duration.component';
