+++ src/app/generated-list-wrapper/generated-list-wrapper.component.html
@@ -1,2 +1,4 @@
+<qpac-header></qpac-header>
 <qpac-list-title></qpac-list-title>
+<qpac-weather-display></qpac-weather-display>
 <qpac-generated-list></qpac-generated-list>
