+++ src/app/generated-list-wrapper/generated-list/generated-list.component.html
@@ -1,12 +1,4 @@
 <div class="card" *ngIf="listItems.length">
-  <h2>Lviv</h2>
-  <p>
-    October 23 - November 7
-  </p>
-  <div>
-    Weather block.......
-    <button (click)="showListService()">ListItem</button>
-  </div>
   <div>
     <div>
       <mat-radio-group>
