+++ src/app/generated-list-wrapper/weather-display/weather-display.component.html
@@ -1,0 +1,9 @@
+<div *ngIf="response | async as responses">
+  <div *ngFor="let attr of responses">
+    <ul>
+      <li> {{ 'weather.display.temperature.text' | translate }} {{ attr.currently.temperature | json }} &#8451; </li>
+      <li> {{ 'weather.display.summary.text' | translate}} {{ attr.currently.summary | json }} </li>
+      <li> {{ 'weather.display.location.text' | translate}} {{ attr.timezone | json }} </li>
+    </ul>
+  </div>
+</div>
